requirements.txt
django 
djangorestframework
pyyaml
requests
django-cors-headers
python-dotenv
gunicorn
whitenoise
dj-database-url
psycopg2
django-filter
djangorestframework-simplejwt